Making all GL operations happen on the queue thread. This prevents us from requiring multiple GL contexts and makes GL use much easier to debug. With this set of changes we now create 3 contexts in normal cases - one for factory/querying, one for each gfx::Context, and one for each EGL gfx::SwapChain backed by a native surface. The factory context could potentially be optimized out on EGL (but not WGL).
This is accomplished with the ES3QueueObject interface that allows objects
to describe what operations should run with a GL context and a mechanism to
marshal those operations to the queue that owns that context.

GLSL shader compilation/program linking now happen on the queue threads,
which is nice for main thread perf but obscures errors. Will need to think
of a better solution to this in the long term (besides CHECKing).

There's a lot of room for efficiency improvements here particularly around
ES3QueueFence's signal/query, though since these are implementation details
specific to ES3 and not problems with the gfx API they're ok for now.

Future work remains around defining better primitives on Buffer and Image
for dynamic upload/download, potentially with a dedicated ES3 transfer
queue. With this CL they become blocking queue ops and map/unmap is
particularly bad.